+++ Test/Unit/DiscountTest.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Test\Unit;
+
+require_once dirname(__DIR__).'/Setup.php';
+
+use Test\Setup;
+use Braintree;
+
+class DiscountTest extends Setup
+{
+    public function testFactory()
+    {
+        $discount = Braintree\Discount::factory(array());
+
+        $this->assertInstanceOf('Braintree\Discount', $discount);
+    }
+}