+++ src/Braintree/PaymentInstrumentType.php
@@ -4,8 +4,9 @@ namespace Braintree;
 
 final class PaymentInstrumentType
 {
-    const PAYPAL_ACCOUNT      = 'paypal_account';
-    const COINBASE_ACCOUNT      = 'coinbase_account';
-    const SEPA_BANK_ACCOUNT   = 'sepa_bank_account';
+     const PAYPAL_ACCOUNT      = 'paypal_account';
+    const COINBASE_ACCOUNT    = 'coinbase_account';
+    const EUROPE_BANK_ACCOUNT = 'europe_bank_account';
     const CREDIT_CARD         = 'credit_card';
+    const APPLE_PAY_CARD      = 'apple_pay_card';
 }